Modal and stuff
There are quite some things. Don't want to go over them all. Most important is that there is a modal that will pop up to fill in a form to create a new game.

I'm also thinking about how to go around the whole log in system for easier development.