Test two AI's playing to each other
Created a test that make two AI's play to each other through the whole lobby logic.
Rewrote a whole lot for this. The lobby now knows nothing about AI's, that abstracted away from it.
AI's know when to play by triggering an event they are waiting for based on what it gets as message.
Also changes a lot of awaits to tasks to prevent any blockings