Choose suit in client
The frontend will ask the player to select the suit the next card should have using a modal. All game logic is adjusted to this. The chosen suit is half opaque shown on the playdeck. The state is all in the server so the client can refresh the page whenever.